Add chart to ChantDetails
